requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
fake-useragent==0.1.11
idna==2.8
lxml==4.3.2
numpy==1.16.2
Pillow==5.4.1
PyAutoGUI==0.9.42
PyGetWindow==0.0.4
PyMsgBox==1.0.6
PyRect==0.1.4
PyScreeze==0.1.20
PySide2==5.12.2
PyTweening==1.0.3
requests==2.21.0
scipy==1.2.1
selenium==3.141.0
shiboken2==5.12.2
sip==4.19.8
soupsieve==1.9
urllib3==1.24.1
